rename art decorator
